requirements.txt

---

## 🧰 **requirements.txt 🧰 **
```txt
pandas
numpy
matplotlib.pyplot
seaborn
scikit-learn
opacus
sdv
faker
scipy
tqdm
